refactor: extract shared 3D viewer operations
